GigaBot/hardware-monitor/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.31.0
redis==5.0.1
